+++ mirage/factories/contributor.ts
@@ -9,7 +9,15 @@ export default Factory.extend<Contributor>({
     unregisteredContributor() {
         return faker.random.number(5) ? undefined : faker.name.firstName();
     },
-    index: 0,
+    index(i: number) {
+        return i;
+    },
     node: association() as Contributor['node'],
     users: association() as Contributor['users'],
 });
+
+declare module 'ember-cli-mirage/types/registries/schema' {
+    export default interface MirageSchemaRegistry {
+        contributors: Contributor;
+    } // eslint-disable-line semi
+}