+++ mirage/views/update-email.ts
@@ -1,14 +1,11 @@
 import { HandlerContext, Request, Schema } from 'ember-cli-mirage';
 
-import User from 'ember-osf-web/models/user';
-import UserEmail from 'ember-osf-web/models/user-email';
-
 export function updateEmails(this: HandlerContext, schema: Schema, request: Request) {
     const { parentID } = request.params;
     const { emailID } = request.params;
-    const emailList = schema.users.find<User>(parentID).emails;
-    const attrs = this.normalizedRequestAttrs('userEmail');
-    const updatedInfo = schema.userEmails.find<UserEmail>(emailID).update(attrs);
+    const emailList = schema.users.find(parentID).emails;
+    const attrs = this.normalizedRequestAttrs('user-email');
+    const updatedInfo = schema.userEmails.find(emailID).update(attrs);
 
     /*
     // If the user is updating an email to be primary,
@@ -29,8 +26,8 @@ export function updateEmails(this: HandlerContext, schema: Schema, request: Requ
 
 export function createEmails(this: HandlerContext, schema: Schema, request: Request) {
     const { parentID } = request.params;
-    const attrs = this.normalizedRequestAttrs('userEmail');
-    const user = schema.users.find<User>(parentID);
+    const attrs = this.normalizedRequestAttrs('user-email');
+    const user = schema.users.find(parentID);
     /*
     // If the user is adding an email, check if primary.
     // If has primary value use that. Otherwise it's false.